Marc Parser
Generating Jsons out of the data
"Reads the Marc File and identifies the lines that belong to the same
book. It generates a dictionary of a book and adds to a general
dictionary. When the next book is generated, then it either added to all
the books list or merges with previous enteries of the same book.